requirements.txt
django~=4.2.8
pillow~=10.2.0
djangorestframework~=3.14.0
python-dotenv~=1.0.0
django-cors-headers~=4.3.1
drf-spectacular~=0.27.0
djangorestframework-simplejwt~=5.3.1